+++ CMake/CatchConfigOptions.cmake
@@ -26,6 +26,7 @@ endmacro()
 
 set(_OverridableOptions
   "ANDROID_LOGWRITE"
+  "COLOUR_WIN32"
   "COUNTER"
   "CPP11_TO_STRING"
   "CPP17_BYTE"
@@ -45,9 +46,6 @@ foreach(OptionName ${_OverridableOptions})
 endforeach()
 
 set(_OtherConfigOptions
-  "COLOUR_ANSI"
-  "COLOUR_NONE"
-  "COLOUR_WINDOWS"
   "DISABLE_EXCEPTIONS"
   "DISABLE_EXCEPTIONS_CUSTOM_HANDLER"
   "DISABLE"